Flytter hentBarnMedLøpendeUtbetaling til fagsak.ts slik at det blir likt som i BA
