+++ src/frontend/utils/fagsak.ts
@@ -1,3 +1,6 @@
+import { isAfter } from 'date-fns';
+
+import { dagensDato, isoStringTilDateMedFallback, tidenesEnde } from './dato';
 import type { VisningBehandling } from '../sider/Fagsak/Saksoversikt/visningBehandling';
 import type { IMinimalFagsak } from '../typer/fagsak';
 import { fagsakStatus } from '../typer/fagsak';
@@ -14,3 +17,22 @@ export const hentAktivBehandlingPåMinimalFagsak = (
 ): VisningBehandling | undefined => {
     return minimalFagsak.behandlinger.find((behandling: VisningBehandling) => behandling.aktiv);
 };
+
+export const hentBarnMedLøpendeUtbetaling = (minimalFagsak: IMinimalFagsak): Set<string> =>
+    minimalFagsak.gjeldendeUtbetalingsperioder
+        .filter(utbetalingsperiode =>
+            isAfter(
+                isoStringTilDateMedFallback({
+                    isoString: utbetalingsperiode.periodeTom,
+                    fallbackDate: tidenesEnde,
+                }),
+                dagensDato
+            )
+        )
+        .reduce((acc, utbetalingsperiode) => {
+            utbetalingsperiode.utbetalingsperiodeDetaljer.map(utbetalingsperiodeDetalj =>
+                acc.add(utbetalingsperiodeDetalj.person.personIdent)
+            );
+
+            return acc;
+        }, new Set<string>());
